+++ index.php
@@ -20,6 +20,7 @@ $sitename      = htmlspecialchars($app->get('sitename'), ENT_QUOTES, 'UTF-8');
 $pageclass     = $app->getMenu()->getActive()->getParams()->get('pageclass_sfx');
 $themeSwitcher = (boolean)$this->params->get('theme-switcher', 1);
 $fontAwesome   = (boolean)$this->params->get('font-awesome-thats-actually-rather-shit', 1);
+$googleFont    = $this->params->get('google-font', '');
 
 // Load switcher CSS
 if ($themeSwitcher)
@@ -34,6 +35,19 @@ $wa->usePreset('template.lightning');
 if ($fontAwesome)
 {
 	$wa->useStyle('fontawesome.css');
+}
+
+// Google font
+if ($googleFont !== '')
+{
+	$fontFamily = str_replace(' ', '+', $googleFont);
+	$this->getPreloadManager()->preconnect('https://fonts.googleapis.com', []);
+	$this->getPreloadManager()->preconnect('https://fonts.gstatic.com', []);
+	$this->addHeadLink('https://fonts.googleapis.com/css2?family=' . $fontFamily . '&display=swap', 'stylesheet', 'rel');
+
+	$wa->addInlineStyle(':root {
+		--hiq-font-family-base: "' . $googleFont . '";
+	}');
 }
 
 // Load switcher JS